parkingSpotterBackend/requirements.txt
flask==3.0.2
flask-socketio==5.3.6
flask-cors==4.0.0
waitress==2.1.2
python-socketio==5.11.1
psycopg2==2.9.9  # Changed from psycopg2-binary
SQLAlchemy==2.0.27
alembic==1.13.1
requests==2.31.0
pillow==10.2.0
python-dotenv==1.0.1
haversine==2.8.0
psutil==5.9.8
googlemaps==4.10.0
Werkzeug==3.0.1  # Required for secure file handling
typing-extensions==4.9.0  # Required for type hints
flask-limiter==3.5.0  # Rate limiting
flask-talisman==1.1.0  # HTTPS enforcement 
